+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   > underlying rest5client is using apache http async client + jackson 3.0
 * json: updated jackson api to 3.1.0
   > major jackson api update, now deserialization throws JSONException if json is invalid, instead of UncheckedIOException
+* monitor: improve kafka disk monitor error message
 
 ### 9.4.2 (1/26/26 - 2/15/26)